feat: update profile page content
