+++ app/routes/auth/profileStyle.css
@@ -1,0 +1,16 @@
+.profileHeader,
+.profileContent {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.profileHeader {
+  font-weight: bold;
+  font-size: xx-large;
+  margin: 0.5rem 0;
+}
+
+.profileInfo {
+  padding: 1rem;
+}
